--- src/cli/read-robinhood-pdf/types.ts
@@ -1,22 +1,0 @@
-export interface TextContent {
-  items: TextItem[];
-}
-
-export interface TextItem {
-  str: string;
-  transform: number[];
-  width: number;
-  height: number;
-}
-
-export enum PageType {
-  AccountActivity = 'ACCOUNT ACTIVITY',
-  PortfolioSummary = 'PORTFOLIO SUMMARY',
-  Unknown = 'UNKNOWN'
-}
-
-export interface Column {
-  text: string;
-  startX: number;
-  // items: TextItem[];
-}
